# Optimize newChunk()