Some reorganization can somewhat improve the performance of nextChunk(). This also makes function min() redundant.

```
name     old time/op   new time/op   delta
FastCDC    1.85s ± 0%    1.68s ± 1%  -8.99%  (p=0.000 n=10+10)

name     old speed     new speed     delta
FastCDC  567MB/s ± 0%  623MB/s ± 1%  +9.87%  (p=0.000 n=10+10)
```

Signed-off-by: Grzegorz Kowal <grzegorz@amuncode.org>